Refactor Landing page to handle undefined, null, and valid sessions.
