+++ src/components/Landing/MessageOfTheDay.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+export default function() {
+  return (
+    <p>
+      Hello! This is a tool for collaboration. It is a WebDevAtlanta project.{" "}
+      <a href="https://github.com/webdevatlanta/GroupCollaborationTool">
+        See the code here
+      </a>
+    </p>
+  );
+}
